Allocate surface IDs on client side
This is much easier to manage for clients using surfaces in an
asynchronous way (i.e. over a mojo pipe or chromium IPC). The code
translating from the IPC layer to SurfaceFactory can validate that the
namespace of SurfaceIds matches the logical connection it is managing.

Review URL: https://codereview.chromium.org/358003003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280950 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: cde7921316422aad0569127571623afd099df5f1

+++ cc/surfaces/surface_id.h
@@ -12,11 +12,11 @@ namespace cc {
 
 struct SurfaceId {
   SurfaceId() : id(0) {}
-  explicit SurfaceId(int id) : id(id) {}
+  explicit SurfaceId(uint64_t id) : id(id) {}
 
   bool is_null() const { return id == 0; }
 
-  int id;
+  uint64_t id;
 };
 
 inline bool operator==(const SurfaceId& a, const SurfaceId& b) {
@@ -41,7 +41,9 @@ inline size_t hash_value(const cc::SurfaceId& key) {
 #elif defined(COMPILER_GCC)
 template <>
 struct hash<cc::SurfaceId> {
-  size_t operator()(cc::SurfaceId key) const { return hash<int>()(key.id); }
+  size_t operator()(cc::SurfaceId key) const {
+    return hash<uint64_t>()(key.id);
+  }
 };
 #else
 #error define a hash function for your compiler
